+++ jupiter-registry/jupiter-registry-default/src/main/java/org/jupiter/registry/DefaultRegistry.java
@@ -452,10 +452,8 @@ public final class DefaultRegistry extends NettyTcpConnector {
                         ch.writeAndFlush(new Acknowledge(obj.sequence()))  // 回复ACK
                                 .addListener(ChannelFutureListener.FIRE_EXCEPTION_ON_FAILURE);
 
-                        if (logger.isInfoEnabled()) {
-                            logger.info("Publish from RegistryServer {}, metadata: {}, version: {}.",
-                                    data.getFirst(), metaObj, obj.version());
-                        }
+                        logger.info("Publish from RegistryServer {}, metadata: {}, version: {}.",
+                                data.getFirst(), metaObj, obj.version());
 
                         break;
                     }
@@ -468,10 +466,8 @@ public final class DefaultRegistry extends NettyTcpConnector {
                         ch.writeAndFlush(new Acknowledge(obj.sequence()))  // 回复ACK
                                 .addListener(ChannelFutureListener.FIRE_EXCEPTION_ON_FAILURE);
 
-                        if (logger.isInfoEnabled()) {
-                            logger.info("Publish cancel from RegistryServer {}, metadata: {}, version: {}.",
-                                    data.getFirst(), data.getSecond(), obj.version());
-                        }
+                        logger.info("Publish cancel from RegistryServer {}, metadata: {}, version: {}.",
+                                data.getFirst(), data.getSecond(), obj.version());
 
                         break;
                     }
@@ -487,9 +483,7 @@ public final class DefaultRegistry extends NettyTcpConnector {
             } else if (msg instanceof Acknowledge) {
                 handleAcknowledge((Acknowledge) msg);
             } else {
-                if (logger.isWarnEnabled()) {
-                    logger.warn("Unexpected message type received: {}, channel: {}.", msg.getClass(), ch);
-                }
+                logger.warn("Unexpected message type received: {}, channel: {}.", msg.getClass(), ch);
 
                 ReferenceCountUtil.release(msg);
             }