+++ jupiter-rpc/src/main/java/org/jupiter/rpc/DefaultClient.java
@@ -83,7 +83,9 @@ public class DefaultClient implements JClient {
 
     @Override
     public JClient withConnector(JConnector<JConnection> connector) {
-        connector.withProcessor(new DefaultConsumerProcessor());
+        if (connector.processor() == null) {
+            connector.withProcessor(new DefaultConsumerProcessor());
+        }
         this.connector = connector;
         return this;
     }
@@ -237,6 +239,17 @@ public class DefaultClient implements JClient {
     }
 
     @Override
+    public boolean awaitConnections(Class<?> interfaceClass, long timeoutMillis) {
+        return awaitConnections(interfaceClass, JConstants.DEFAULT_VERSION, timeoutMillis);
+    }
+
+    @Override
+    public boolean awaitConnections(Class<?> interfaceClass, String version, long timeoutMillis) {
+        JConnector.ConnectionWatcher watcher = watchConnections(interfaceClass, version);
+        return watcher.waitForAvailable(timeoutMillis);
+    }
+
+    @Override
     public boolean awaitConnections(Directory directory, long timeoutMillis) {
         JConnector.ConnectionWatcher watcher = watchConnections(directory);
         return watcher.waitForAvailable(timeoutMillis);
@@ -258,6 +271,7 @@ public class DefaultClient implements JClient {
 
     @Override
     public void shutdownGracefully() {
+        registryService.shutdownGracefully();
         connector.shutdownGracefully();
     }