+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/future/DefaultInvokeFuture.java
@@ -21,15 +21,16 @@ import org.jupiter.common.util.Maps;
 import org.jupiter.common.util.Signal;
 import org.jupiter.common.util.internal.logging.InternalLogger;
 import org.jupiter.common.util.internal.logging.InternalLoggerFactory;
-import org.jupiter.rpc.ConsumerHook;
 import org.jupiter.rpc.DispatchType;
 import org.jupiter.rpc.JListener;
 import org.jupiter.rpc.JResponse;
+import org.jupiter.rpc.consumer.ConsumerInterceptor;
 import org.jupiter.rpc.exception.JupiterBizException;
 import org.jupiter.rpc.exception.JupiterRemoteException;
 import org.jupiter.rpc.exception.JupiterSerializationException;
 import org.jupiter.rpc.exception.JupiterTimeoutException;
 import org.jupiter.rpc.model.metadata.ResultWrapper;
+import org.jupiter.rpc.tracing.TraceId;
 import org.jupiter.transport.Status;
 import org.jupiter.transport.channel.JChannel;
 
@@ -37,7 +38,6 @@ import java.net.SocketAddress;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.TimeUnit;
 
-import static org.jupiter.common.util.Preconditions.checkNotNull;
 import static org.jupiter.common.util.StackTraceUtil.stackTrace;
 
 /**
@@ -63,7 +63,8 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
 
     private volatile boolean sent = false;
 
-    private ConsumerHook[] hooks = ConsumerHook.EMPTY_HOOKS;
+    private ConsumerInterceptor[] interceptors;
+    private TraceId traceId;
 
     public static <T> DefaultInvokeFuture<T> with(
             long invokeId, JChannel channel, Class<T> returnType, long timeoutMillis, DispatchType dispatchType) {
@@ -129,14 +130,21 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
         sent = true;
     }
 
-    public ConsumerHook[] hooks() {
-        return hooks;
+    public ConsumerInterceptor[] interceptors() {
+        return interceptors;
     }
 
-    public DefaultInvokeFuture<V> hooks(ConsumerHook[] hooks) {
-        checkNotNull(hooks, "hooks");
+    public DefaultInvokeFuture<V> interceptors(ConsumerInterceptor[] interceptors) {
+        this.interceptors = interceptors;
+        return this;
+    }
 
-        this.hooks = hooks;
+    public TraceId traceId() {
+        return traceId;
+    }
+
+    public DefaultInvokeFuture<V> traceId(TraceId traceId) {
+        this.traceId = traceId;
         return this;
     }
 
@@ -151,9 +159,11 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
             setException(status, response);
         }
 
-        // call hook's after method
-        for (int i = 0; i < hooks.length; i++) {
-            hooks[i].after(response, channel);
+        ConsumerInterceptor[] interceptors = this.interceptors; // snapshot
+        if (interceptors != null) {
+            for (int i = interceptors.length - 1; i >= 0; i--) {
+                interceptors[i].afterInvoke(traceId, response, channel);
+            }
         }
     }
 
@@ -187,14 +197,35 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
 
     public static void received(JChannel channel, JResponse response) {
         long invokeId = response.id();
+
         DefaultInvokeFuture<?> future = roundFutures.remove(invokeId);
+
         if (future == null) {
             // 广播场景下做出了一点让步, 多查询了一次roundFutures
             future = broadcastFutures.remove(subInvokeId(channel, invokeId));
         }
+
         if (future == null) {
             logger.warn("A timeout response [{}] finally returned on {}.", response, channel);
             return;
+        }
+
+        future.doReceived(response);
+    }
+
+    public static void fakeReceived(JChannel channel, JResponse response, DispatchType dispatchType) {
+        long invokeId = response.id();
+
+        DefaultInvokeFuture<?> future = null;
+
+        if (dispatchType == DispatchType.ROUND) {
+            future = roundFutures.remove(invokeId);
+        } else if (dispatchType == DispatchType.BROADCAST) {
+            future = broadcastFutures.remove(subInvokeId(channel, invokeId));
+        }
+
+        if (future == null) {
+            return; // 正确结果在超时被处理之前返回
         }
 
         future.doReceived(response);
@@ -213,12 +244,12 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
                 try {
                     // round
                     for (DefaultInvokeFuture<?> future : roundFutures.values()) {
-                        process(future);
+                        process(future, DispatchType.ROUND);
                     }
 
                     // broadcast
                     for (DefaultInvokeFuture<?> future : broadcastFutures.values()) {
-                        process(future);
+                        process(future, DispatchType.BROADCAST);
                     }
                 } catch (Throwable t) {
                     logger.error("An exception was caught while scanning the timeout futures {}.", stackTrace(t));
@@ -230,7 +261,7 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
             }
         }
 
-        private void process(DefaultInvokeFuture<?> future) {
+        private void process(DefaultInvokeFuture<?> future, DispatchType dispatchType) {
             if (future == null || future.isDone()) {
                 return;
             }
@@ -239,7 +270,7 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
                 JResponse response = new JResponse(future.invokeId);
                 response.status(future.sent ? Status.SERVER_TIMEOUT : Status.CLIENT_TIMEOUT);
 
-                DefaultInvokeFuture.received(future.channel, response);
+                DefaultInvokeFuture.fakeReceived(future.channel, response, dispatchType);
             }
         }
     }