+++ jupiter-transport/jupiter-transport-api/src/main/java/org/jupiter/transport/processor/ProviderProcessor.java
@@ -39,4 +39,6 @@ public interface ProviderProcessor {
      * 处理异常
      */
     void handleException(JChannel channel, JRequestBytes request, Status status, Throwable cause);
+
+    void shutdown();
 }